+++ packages/theme-preset/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @marigold/theme-preset
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @marigold/system@6.9.1
+
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @marigold/system@6.9.0
+
 ## 1.2.1
 
 ### Patch Changes